Read CCAL Data and convert to machine-readable format.

Read CCAL Data and convert to machine-readable format.

Usage

read_ccal(files, concat = FALSE)

Arguments

files

Path to .xlsx file delivered by CCAL. Use a character vector to specify multiple files.

concat

If concat is set to TRUE, the output contains one table for data, one for metadata, one for samples, and one for questionable, rather than one of each table for every input file. By default, concat is set to FALSE, so the output contains separate tables for each file. If only one file path is supplied to the files argument, this parameter does not affect the output.

Value

A nested list. Each list item corresponds to one input file and contains data frames for data and metadata.

Examples

tidy_ccal <- read_ccal(use_example_data(file_names = "SPAC_080199.xlsx"))
#> Reading data from /home/runner/work/_temp/Library/imdccal/extdata/SPAC_080199.x…
#> Warning: ! Questionable results contain duplicate rows:
#> ! See samples: 11: TDN, 13: TDN, and 3: TDN.
#> ℹ This may be caused by lab error, or may be a result of one analyte being
#>   compared to more than one other analyte.
#> ℹ See the `questionable` table returned by this function.