Write data to xlsx or csv file.

Write data to xlsx or csv file.

Usage

write_ccal(
  all_data,
  format = c("xlsx", "csv"),
  destination_folder,
  overwrite,
  suffix,
  num_tables
)

Arguments

all_data

The data to write to file.

format

File format to export machine readable data to - either "xlsx" or "csv"

destination_folder

Folder to save the data in. Defaults to current working directory. Folder must already exist.

overwrite

Overwrite existing file (Defaults to TRUE as with write.table)

suffix

Suffix to add to output file name.

num_tables

Number of tables to write to file.

Examples

if (FALSE) { # \dontrun{
# Create tidied CCAL data from demo data stored in the imdccal package
tidy_ccal <- read_ccal(use_example_data(file_names = "SPAC_080199.xlsx"))

# Write data stored in environment to file
write_ccal(all_data = tidy_ccal,
          format = "xlsx", # alternatively, "csv"
          destination_folder = "ccal_tidy", # must already exist
          overwrite = TRUE,
          suffix = "_tidy",
          num_tables = 4)
} # }